requirements.txt
certifi==2020.12.5
wincertstore==0.2
discord.py==1.6.0
requests==2.23.0
python-dotenv==0.15.0